@@ -18,16 +18,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0"
 
 dependencies:
   carousel_slider: ^4.1.1
+  cloud_firestore: ^3.5.0
+  email_validator: ^2.1.17
+  firebase_auth: ^3.11.1
   flutter:
     sdk: flutter
   flutter_svg: ^1.1.5
   get:
   google_fonts: ^3.0.1
-  shared_preferences: ^2.0.15
 
 dev_dependencies:
   flutter_test:
