@@ -4,7 +4,7 @@ import 'sidebarContainer.dart';
 
 class Search extends StatelessWidget {
   const Search({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
